# backend/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "GenerateScheduleRequest": {
        "additionalProperties": false,
        "properties": {
          "ends_on": {
            "format": "date",
            "title": "Ends On",
            "type": "string"
          },
          "starts_on": {
            "format": "date",
            "title": "Starts On",
            "type": "string"
          }
        },
        "required": [
          "starts_on",
          "ends_on"
        ],
        "title": "GenerateScheduleRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RotationMode": {
        "enum": [
          "hybrid",
          "daily",
          "weekly"
        ],
        "title": "RotationMode",
        "type": "string"
      },
...
      "ScheduleSummaryResponse": {
        "description": "A schedule without its assignments, for the drafts listing.",
        "properties": {
          "assignment_count": {
            "title": "Assignment Count",
            "type": "integer"
          },
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "ends_on": {
            "format": "date",
            "title": "Ends On",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "rotation_mode": {
            "$ref": "#/components/schemas/RotationMode"
          },
          "solver_status": {
            "title": "Solver Status",
            "type": "string"
          },
          "starts_on": {
            "format": "date",
            "title": "Starts On",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name",
          "starts_on",
          "ends_on",
          "status",
          "version",
          "rotation_mode",
          "solver_status",
          "assignment_count",
          "created_at"
        ],
        "title": "ScheduleSummaryResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Erste On-call",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/scheduling/compare": {
      "get": {
        "operationId": "compare_schedules_api_v1_scheduling_compare_get",
        "parameters": [
          {
            "in": "query",
            "name": "left_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Left Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "right_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Right Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "oncall_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Oncall Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Compare Schedules Api V1 Scheduling Compare Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Compare Schedules",
        "tags": [
          "scheduling"
        ]
      }
    },
    "/api/v1/scheduling/drafts": {
      "get": {
        "description": "Drafts and proposals still awaiting a decision.\n\nWithout this the generator result lived only in component state: reloading\nthe page orphaned the draft in the database with no way back to it.",
        "operationId": "list_drafts_api_v1_scheduling_drafts_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "cookie",
            "name": "oncall_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Oncall Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ScheduleSummaryResponse"
                  },
                  "title": "Response List Drafts Api V1 Scheduling Drafts Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Drafts",
        "tags": [
          "scheduling"
        ]
      }
    },
    "/api/v1/scheduling/policy": {
      "get": {
        "operationId": "get_policy_api_v1_scheduling_policy_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "oncall_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Oncall Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SchedulingPolicyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Policy",
        "tags": [
          "scheduling"
        ]
      },
      "put": {
        "operationId": "update_policy_api_v1_scheduling_policy_put",
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "cookie",
            "name": "oncall_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Oncall Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SchedulingPolicyUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SchedulingPolicyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Policy",
        "tags": [
          "scheduling"
        ]
      }
    },
    "/api/v1/scheduling/runs": {
      "get": {
        "description": "This coordinator's generations that are still in flight.\n\nA reload used to lose the running job even though the database had it, and\nthe natural reaction to a screen that shows nothing is to start a second\ngeneration - which leaves two indistinguishable drafts (MED5-11, LOW5-09).",
        "operationId": "list_runs_api_v1_scheduling_runs_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "cookie",
            "name": "oncall_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Oncall Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response List Runs Api V1 Scheduling Runs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Runs",
        "tags": [
          "scheduling"
        ]
      },
      "post": {
        "operationId": "queue_generation_api_v1_scheduling_runs_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "cookie",
            "name": "oncall_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Oncall Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateScheduleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Queue Generation Api V1 Scheduling Runs Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Queue Generation",
        "tags": [
          "scheduling"
        ]
      }
    },
    "/api/v1/scheduling/runs/{run_id}": {
      "get": {
        "operationId": "generation_status_api_v1_scheduling_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "oncall_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Oncall Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Generation Status Api V1 Scheduling Runs  Run Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generation Status",
        "tags": [
          "scheduling"
        ]
      }
    },
    "/api/v1/scheduling/suggested-range": {
      "get": {
        "operationId": "suggested_range_api_v1_scheduling_suggested_range_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "oncall_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Oncall Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "format": "date",
                    "type": "string"
                  },
                  "title": "Response Suggested Range Api V1 Scheduling Suggested Range Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Suggested Range",
        "tags": [
          "scheduling"
        ]
      }
    },
    "/api/v1/scheduling/{schedule_id}": {
      "delete": {
        "description": "Discard a draft or a proposal.\n\nPublished and superseded schedules are the record of who was on duty, so\nthey are never deletable; abandoned drafts otherwise pile up in the listing.",
        "operationId": "delete_schedule_api_v1_scheduling__schedule_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "schedule_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Schedule Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "cookie",
            "name": "oncall_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Oncall Session"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Schedule",
        "tags": [
          "scheduling"
        ]
      },
      "get": {
        "description": "One schedule with its assignments, so a draft survives a page reload.",
        "operationId": "get_schedule_api_v1_scheduling__schedule_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "schedule_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Schedule Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "oncall_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Oncall Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftScheduleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Schedule",
        "tags": [
          "scheduling"
        ]
      }
    },
    "/api/v1/scheduling/{schedule_id}/fairness-impact": {
      "get": {
        "operationId": "draft_fairness_impact_api_v1_scheduling__schedule_id__fairness_impact_get",
        "parameters": [
          {
            "in": "path",
            "name": "schedule_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Schedule Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "oncall_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Oncall Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftFairnessImpactResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Draft Fairness Impact",
        "tags": [
          "scheduling"
        ]
      }
    },
...
  }
}